requirements.txt
python-dotenv
gradio
openai
pydantic